+++ markdown/tool/stats.dart
@@ -2,19 +2,12 @@ import 'dart:async';
 import 'dart:collection';
 import 'dart:convert';
 import 'dart:io';
-import 'dart:mirrors';
 
 import 'package:args/args.dart';
 import 'package:collection/collection.dart';
-import 'package:html/dom.dart';
-import 'package:html/parser.dart' show parseFragment;
-import 'package:markdown/markdown.dart' show markdownToHtml, ExtensionSet;
 import 'package:path/path.dart' as p;
 
-// Locate the "tool" directory. Use mirrors so that this works with the test
-// package, which loads this suite into an isolate.
-String get _currentDir => p
-    .dirname((reflect(main) as ClosureMirror).function.location.sourceUri.path);
+import 'stats_lib.dart';
 
 Future main(List<String> args) async {
   final parser = new ArgParser()
@@ -24,12 +17,19 @@ Future main(List<String> args) async {
         defaultsTo: false, help: 'raw JSON format', negatable: false)
     ..addFlag('update-files',
         defaultsTo: false,
-        help: 'Update stats files in $_currentDir',
+        help: 'Update stats files in $toolDir',
         negatable: false)
     ..addFlag('verbose',
-        defaultsTo: false, help: 'verbose output', negatable: false)
+        defaultsTo: false,
+        help: 'Print details for failures and errors.',
+        negatable: false)
+    ..addFlag('verbose-loose',
+        defaultsTo: false,
+        help: 'Print details for "loose" matches.',
+        negatable: false)
     ..addOption('flavor',
-        allowed: ['common_mark', 'gfm'], defaultsTo: 'common_mark')
+        allowed: [Config.commonMarkConfig.prefix, Config.gfmConfig.prefix],
+        defaultsTo: Config.commonMarkConfig.prefix)
     ..addFlag('help', defaultsTo: false, negatable: false);
 
   ArgResults options;
@@ -43,7 +43,7 @@ Future main(List<String> args) async {
     return;
   }
 
-  if (options['help']) {
+  if (options['help'] as bool) {
     print(parser.usage);
     return;
   }
@@ -51,6 +51,7 @@ Future main(List<String> args) async {
   var specifiedSection = options['section'] as String;
   var raw = options['raw'] as bool;
   var verbose = options['verbose'] as bool;
+  var verboseLooseMatch = options['verbose-loose'] as bool;
   var updateFiles = options['update-files'] as bool;
 
   if (updateFiles && (raw || verbose || (specifiedSection != null))) {
@@ -60,14 +61,21 @@ Future main(List<String> args) async {
     return;
   }
 
-  final testPrefix = options['flavor'];
+  final testPrefix = options['flavor'] as String;
 
-  ExtensionSet extensionSet;
-  if (testPrefix == 'gfm') {
-    extensionSet = ExtensionSet.gitHub;
+  Config config;
+  switch (testPrefix) {
+    case 'gfm':
+      config = Config.gfmConfig;
+      break;
+    case 'common_mark':
+      config = Config.commonMarkConfig;
+      break;
+    default:
+      throw new ArgumentError('Does not support `$testPrefix`.');
   }
 
-  var sections = _loadCommonMarkSections(testPrefix);
+  var sections = loadCommonMarkSections(testPrefix);
 
   var scores = new SplayTreeMap<String, SplayTreeMap<int, CompareLevel>>(
       compareAsciiLowerCaseNatural);
@@ -80,8 +88,8 @@ Future main(List<String> args) async {
       var nestedMap = scores.putIfAbsent(
           section, () => new SplayTreeMap<int, CompareLevel>());
 
-      nestedMap[e.example] =
-          _compareResult(e, verbose, extensionSet: extensionSet);
+      nestedMap[e.example] = compareResult(config, e,
+          verboseFail: verbose, verboseLooseMatch: verboseLooseMatch);
     }
   });
 
@@ -93,53 +101,6 @@ Future main(List<String> args) async {
     await _printFriendly(testPrefix, scores, updateFiles);
   }
 }
-
-CompareLevel _compareResult(CommonMarkTestCase expected, bool verboseFail,
-    {ExtensionSet extensionSet}) {
-  String output;
-  try {
-    output = markdownToHtml(expected.markdown, extensionSet: extensionSet);
-  } catch (err, stackTrace) {
-    if (verboseFail) {
-      printVerboseFailure(
-          'ERROR', expected, expected.html, 'Thrown: $err\n$stackTrace');
-    }
-
-    return CompareLevel.error;
-  }
-
-  if (expected.html == output) {
-    return CompareLevel.strict;
-  }
-
-  var expectedParsed = parseFragment(expected.html);
-  var actual = parseFragment(output);
-
-  var looseMatch = _compareHtml(expectedParsed.children, actual.children);
-
-  if (!looseMatch && verboseFail) {
-    printVerboseFailure(
-        'FAIL', expected, expectedParsed.outerHtml, actual.outerHtml);
-  }
-
-  return looseMatch ? CompareLevel.loose : CompareLevel.fail;
-}
-
-String indent(String s) => s.splitMapJoin('\n', onNonMatch: (n) => '    $n');
-
-void printVerboseFailure(
-    String message, CommonMarkTestCase test, String expected, String actual) {
-  print('$message: http://spec.commonmark.org/0.27/#example-${test.example}');
-  print('input:');
-  print(indent(test.markdown));
-  print('expected:');
-  print(indent(expected));
-  print('actual:');
-  print(indent(actual));
-  print('-----------------------');
-}
-
-enum CompareLevel { strict, loose, fail, error }
 
 Object _convert(obj) {
   if (obj is CompareLevel) {
@@ -170,9 +131,8 @@ Object _convert(obj) {
 Future _printRaw(String testPrefix, Map scores, bool updateFiles) async {
   IOSink sink;
   if (updateFiles) {
-    var path = p.join(_currentDir, '${testPrefix}_stats.json');
-    print('Updating $path');
-    var file = new File(path);
+    var file = getStatsFile(testPrefix);
+    print('Updating ${file.path}');
     sink = file.openWrite();
   } else {
     sink = stdout;
@@ -202,7 +162,7 @@ Future _printFriendly(
 
   IOSink sink;
   if (updateFiles) {
-    var path = p.join(_currentDir, '${testPrefix}_stats.txt');
+    var path = p.join(toolDir, '${testPrefix}_stats.txt');
     print('Updating $path');
     var file = new File(path);
     sink = file.openWrite();
@@ -228,106 +188,15 @@ Future _printFriendly(
 
     sink.writeln('${sectionValidCount.toString().padLeft(countWidth)} '
         'of ${total.toString().padLeft(countWidth)} '
-        '– ${pct}%  $section');
+        '– $pct%  $section');
   });
 
   var pct = (100 * totalValid / totalExamples).toStringAsFixed(1).padLeft(5);
 
   sink.writeln('${totalValid.toString().padLeft(countWidth)} '
       'of ${totalExamples.toString().padLeft(countWidth)} '
-      '– ${pct}%  TOTAL');
+      '– $pct%  TOTAL');
 
   await sink.flush();
   await sink.close();
-}
-
-/// Compare two DOM trees for equality.
-bool _compareHtml(
-    List<Element> expectedElements, List<Element> actualElements) {
-  if (expectedElements.length != actualElements.length) {
-    return false;
-  }
-
-  for (var childNum = 0; childNum < expectedElements.length; childNum++) {
-    var expected = expectedElements[childNum];
-    var actual = actualElements[childNum];
-
-    if (expected.runtimeType != actual.runtimeType) {
-      return false;
-    }
-
-    if (expected.localName != actual.localName) {
-      return false;
-    }
-
-    if (expected.attributes.length != actual.attributes.length) {
-      return false;
-    }
-
-    var expectedAttrKeys = expected.attributes.keys.toList();
-    expectedAttrKeys.sort();
-
-    var actualAttrKeys = actual.attributes.keys.toList();
-    actualAttrKeys.sort();
-
-    for (var attrNum = 0; attrNum < actualAttrKeys.length; attrNum++) {
-      var expectedAttrKey = expectedAttrKeys[attrNum];
-      var actualAttrKey = actualAttrKeys[attrNum];
-
-      if (expectedAttrKey != actualAttrKey) {
-        return false;
-      }
-
-      if (expected.attributes[expectedAttrKey] !=
-          actual.attributes[actualAttrKey]) {
-        return false;
-      }
-    }
-
-    var childrenEqual = _compareHtml(expected.children, actual.children);
-
-    if (!childrenEqual) {
-      return false;
-    }
-  }
-
-  return true;
-}
-
-Map<String, List<CommonMarkTestCase>> _loadCommonMarkSections(
-    String testPrefix) {
-  var testFile = new File(p.join(_currentDir, '${testPrefix}_tests.json'));
-  var testsJson = testFile.readAsStringSync();
-
-  var testArray = JSON.decode(testsJson) as List<Map<String, dynamic>>;
-
-  var sections = new Map<String, List<CommonMarkTestCase>>();
-
-  for (var exampleMap in testArray) {
-    var exampleTest = new CommonMarkTestCase.fromJson(exampleMap);
-
-    var sectionList =
-        sections.putIfAbsent(exampleTest.section, () => <CommonMarkTestCase>[]);
-
-    sectionList.add(exampleTest);
-  }
-
-  return sections;
-}
-
-class CommonMarkTestCase {
-  final String markdown;
-  final String section;
-  final int example;
-  final String html;
-  final int startLine;
-  final int endLine;
-
-  CommonMarkTestCase(this.example, this.section, this.startLine, this.endLine,
-      this.markdown, this.html);
-
-  factory CommonMarkTestCase.fromJson(Map<String, dynamic> json) {
-    return new CommonMarkTestCase(json['example'], json['section'],
-        json['start_line'], json['end_line'], json['markdown'], json['html']);
-  }
 }